pyproject.toml
[project]
name = "quora-duplicate-evaluator"
version = "0.1.0"
description = "A2A agent that evaluates quality of predictions on whether two queries/questions are duplicates"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "datasets>=2.14.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
